*Rename file
*Fix version overlaping

*Update quote font
